Refactor move keyword parser to own extension
To make it clearer where the keywords are getting handled it is moved to
an own extension within keyword_extension.rs.
