+++ rust/nasl-syntax/src/lib.rs
@@ -1,5 +1,6 @@
 mod cursor;
 mod token;
+mod keyword_extension;
 mod operator_precedence_parser;
 mod parser;
 #[cfg(test)]
